# Annotations: Improve get tags query performance